+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 * Upgrade to scala 2.12.
 * Fix connection leak when retrying on responses with chunked bodies.
+* Add io.l5d.http.allSuccessful and io.l5d.h2.allSuccessful response classifiers.
+* Fail on duplicate config file properties instead of silently taking the last
+  value.
 * Expire idle services and clients.
 
 ## 1.0.0 2017-04-24